requirements.txt
# Core ML Libraries
tensorflow==2.15.0
keras==2.15.0
numpy==1.26.4
pandas==2.1.3
scikit-learn==1.3.2

# Image Processing
opencv-python==4.12.0.88
Pillow==11.3.0

# Data Acquisition
kagglehub==0.2.8

# Visualization
matplotlib==3.10.6
seaborn==0.13.2
plotly==5.24.1

# API Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# Web UI
streamlit==1.41.1

# Load Testing
locust==2.32.5

# Utilities
python-dotenv==1.0.0
requests==2.32.5
tqdm==4.67.1

# Jupyter
jupyter==1.1.1
ipykernel==6.30.1
ipywidgets==8.1.7
